+++ plugin-packs/postcss-preset-env/src/test/log/test-logger.mjs
@@ -1,0 +1,14 @@
+export function testLogger() {
+	let logs = [];
+
+	return {
+		warn: (line) => {
+			logs.push(line);
+		},
+		getLogs: () => {
+			const x = logs.slice();
+			logs = [];
+			return x;
+		},
+	};
+}